Allow selectOption() to select options not inside forms.
This fixes a regression in behavior that was introduced in #1468.
It includes a new unit test to ensure that this behavior can be preserved
going forward in future modifications. I preform two sets of checks so that
if the ancestor::form is catching things we don't risk selecting the wrong
element. Confirmed this works for the tests for issue #1467 as well.