Print Selenium logs on failure
WebDriver provides access to the logs for both the client and server, which are sometimes necessary
to debug test failures. This commit modifies Codeception\Module\WebDriver->_failed() to grab all
available logs and print them in human-readable format. The format is identical to what Selenium
Server uses for displaying logs in the console.